PREVIEW: t5516: stop testing .git/branches/ functionality
We are about to remove support for .git/remotes/ and .git/branches/, at
long last. In preparation for this move, let's remove the corresponding
regression tests from t5516-fetch-push.

Note: the `push --prune` test case (and another test case after that)
relied on the side effect where a branch with the name `second` was
created earlier. We simply move that branch creation into that test
case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>