t5003: stop assuming that there is a /bin/sh
This is an easy assumption to make for any Unix old-timer. It's still
an incorrect assumption, though.

In Git for Windows' setup, it sort of worked so far, as the `/bin/`
directory is simply mapped to `/usr/bin/` in Git Bash.

However, when using a different Unix shell (such as BusyBox-w32's ash),
the `/bin/` directory is not mapped. Proving that assumption to be
false.

So let's just use git.exe instead. It's just a binary that we need,
anyway, and in contrast to /bin/sh, we can be certain at this point that
git.exe exists.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>