mingw: make it easier to compile for debugging with gdb
Single-stepping in gdb works a lot better when compiling without -O2
(otherwise gcc might reorder operations to make the code faster, but
that happens to be pretty confusing when trying to pin-point a problem,
not to mention frustrating).

On Windows, additionally there is the problem that gdb cannot set
breakpoints when the code was compiled with Address Space Layout
Randomization (ASLR).

So let's make it easier to compile specifically for debugging with gdb:
`make DEBUG=1` will strip out the appropriate flags so that the code is
compiled without optimization and without ASLR.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>