mingw: introduce core.longpaths = quick
The thing about the \\?\ trick is: it not only works around path length
limitations, it side-steps a whole lot of code in the Win32 API that
imposes other restrictions such as NUL or AUX being prohibited path
components, even with file extensions (AUX.c is illegal on Windows, for
example).

This *should* result in a nice performance win.

However, since we use CreateFile()/GetFinalPathByHandle(), we lose all
that speed again, as CreateFile() goes through the same validation
routines.

The reason we call GetFinalPathByHandle() is to normalize path
components . and .., i.e. hello/world/../you => hello/you (because \\?\
prefixed paths cannot contain . nor .. as path components).

Let's introduce a new value for core.longpaths that triggers hand-rolled
code to normalize the paths.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>