mingw: make use of Windows 10' Win 32 long paths registry settings
As of Creators Update, Windows 10 supports a registry setting that lifts
the MAX_PATH limit of 260.

Let's make use of that feature also in Git, by providing it with the
appropriate manifest adjustment, and adjusting Git's idea of MAX_PATH.

While at it, #include <winuser.h> in git.rc so that RT_MANIFEST is
interpreted correctly (otherwise, the manifest is not included correctly
with Git for Windows' current SDK).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>